그룹 조화수 (서버)쿠키, 멤버아이디 기준으로 중복되지않게 추가
다음작업, 오늘조회수, 좋아요 등 작업예정

+++ src/main/java/com/site/pine/dto/group/GroupContentResDto.java
@@ -22,11 +22,11 @@ public class GroupContentResDto {
     private Integer userLimit;
 
     // reqDto에 없는 값들
-    private Integer allViewCount;
-    private Integer todayViewCount;
-    private Integer likeCount;
-    private Integer postCount;
-    private Integer groupMemberCount;
+    private Long allViewCount;
+    private Long todayViewCount;
+    private Long likeCount;
+    private Long postCount;
+    private Long groupMemberCount;
 
     private Timestamp indate;
